[#36] Removed old progress indicator
