+++ src/components/ProgressIndicator/index.js
@@ -1,30 +1,34 @@
 import PropTypes from 'prop-types';
-import React, {useContext, useEffect, useState} from 'react';
-import {useLocation, useMatch} from 'react-router-dom';
+import React, {useEffect, useState} from 'react';
+import {FormattedMessage} from 'react-intl';
+import {useLocation} from 'react-router-dom';
 
-import {ConfigContext} from 'Context';
-import {SUBMISSION_ALLOWED} from 'components/constants';
-import {IsFormDesigner} from 'headers';
-import Types from 'types';
+import Caption from 'components/Caption';
+import Card from 'components/Card';
+import List from 'components/List';
+import {STEP_LABELS} from 'components/constants';
 
-import ProgressIndicatorDisplay from './ProgressIndicatorDisplay';
-import {STEP_LABELS} from './constants';
+import MobileButton from './MobileButton';
+import ProgressIndicatorItem from './ProgressIndicatorItem';
 
 const ProgressIndicator = ({
-  title,
-  submission = null,
+  progressIndicatorTitle,
+  formTitle,
   steps,
-  submissionAllowed,
-  completed = false,
-  hideNonApplicableSteps = false,
+  ariaMobileIconLabel,
+  accessibleToggleStepsLabel,
+  sticky = true,
 }) => {
-  const {pathname} = useLocation();
-  const config = useContext(ConfigContext);
-  const summaryMatch = !!useMatch('/overzicht');
-  const stepMatch = useMatch('/stap/:step');
-  const confirmationMatch = !!useMatch('/bevestiging');
-  const isStartPage = !summaryMatch && stepMatch == null && !confirmationMatch;
+  const {pathname: currentPathname} = useLocation();
   const [expanded, setExpanded] = useState(false);
+
+  const modifiers = [];
+  if (sticky) {
+    modifiers.push('sticky');
+  }
+  if (!expanded) {
+    modifiers.push('mobile-collapsed');
+  }
 
   // collapse the expanded progress indicator if nav occurred, see
   // open-formulieren/open-forms#2673. It's important that *only* the pathname triggers
@@ -34,109 +38,63 @@ const ProgressIndicator = ({
       setExpanded(false);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pathname]);
-
-  // figure out the slug from the currently active step IF we're looking at a step
-  const stepSlug = stepMatch ? stepMatch.params.step : '';
-  const hasSubmission = !!submission;
-
-  const applicableSteps = hasSubmission ? submission.steps.filter(step => step.isApplicable) : [];
-  const applicableAndCompletedSteps = applicableSteps.filter(step => step.completed);
-  const applicableCompleted =
-    hasSubmission && applicableSteps.length === applicableAndCompletedSteps.length;
-
-  // If any step cannot be submitted, there should NOT be an active link to the overview page.
-  const canSubmitSteps = hasSubmission
-    ? submission.steps.filter(step => !step.canSubmit).length === 0
-    : false;
-
-  // figure out the title for the mobile menu based on the state
-  let activeStepTitle;
-  if (isStartPage) {
-    activeStepTitle = STEP_LABELS.login;
-  } else if (summaryMatch) {
-    activeStepTitle = STEP_LABELS.overview;
-  } else if (confirmationMatch) {
-    activeStepTitle = STEP_LABELS.confirmation;
-  } else {
-    const step = steps.find(step => step.slug === stepSlug);
-    activeStepTitle = step.formDefinition;
-  }
-
-  const canNavigateToStep = index => {
-    // The user can navigate to a step when:
-    // 1. All previous steps have been completed
-    // 2. The user is a form designer
-    if (IsFormDesigner.getValue()) return true;
-
-    if (!submission) return false;
-
-    const previousSteps = submission.steps.slice(0, index);
-    const previousApplicableButNotCompletedSteps = previousSteps.filter(
-      step => step.isApplicable && !step.completed
-    );
-
-    return !previousApplicableButNotCompletedSteps.length;
-  };
-
-  const getStepsInfo = steps => {
-    return steps.map((step, index) => ({
-      uuid: step.uuid,
-      slug: step.slug,
-      to: step.href || `/stap/${step.slug}`,
-      formDefinition: step.formDefinition,
-      isCompleted: submission ? submission.steps[index].completed : false,
-      isApplicable: submission ? submission.steps[index].isApplicable : step.isApplicable ?? true,
-      isCurrent: step.slug === stepSlug,
-      canNavigateTo: canNavigateToStep(index),
-    }));
-  };
-
-  // try to get the value from the submission if provided, otherwise
-  const submissionAllowedSpec = submission?.submissionAllowed ?? submissionAllowed;
-  const showOverview = submissionAllowedSpec !== SUBMISSION_ALLOWED.noWithoutOverview;
-  const showConfirmation = submissionAllowedSpec === SUBMISSION_ALLOWED.yes;
-
-  const ProgressIndicatorDisplayComponent =
-    config?.displayComponents?.progressIndicator ?? ProgressIndicatorDisplay;
+  }, [currentPathname]);
 
   return (
-    <ProgressIndicatorDisplayComponent
-      activeStepTitle={activeStepTitle}
-      formTitle={title}
-      steps={getStepsInfo(steps)}
-      hideNonApplicableSteps={hideNonApplicableSteps}
-      hasSubmission={hasSubmission}
-      isStartPage={isStartPage}
-      isSummary={summaryMatch}
-      isConfirmation={confirmationMatch}
-      isSubmissionComplete={completed}
-      areApplicableStepsCompleted={applicableCompleted && canSubmitSteps}
-      showOverview={showOverview}
-      showConfirmation={showConfirmation}
-      expanded={expanded}
-      onExpandClick={() => setExpanded(!expanded)}
-    />
+    <Card blockClassName="progress-indicator" modifiers={modifiers}>
+      <nav>
+        <MobileButton
+          ariaMobileIconLabel={ariaMobileIconLabel}
+          accessibleToggleStepsLabel={accessibleToggleStepsLabel}
+          formTitle={formTitle}
+          expanded={expanded}
+          onExpandClick={() => setExpanded(!expanded)}
+        />
+        <Caption component="h2">
+          <FormattedMessage
+            description="Title of progress indicator"
+            defaultMessage="{progressIndicatorTitle}"
+            values={{progressIndicatorTitle}}
+          />
+        </Caption>
+        <List ordered>
+          {steps.map(step => (
+            <ProgressIndicatorItem
+              key={step.slug || step.uuid}
+              text={step.formDefinition}
+              href={step.to}
+              isActive={step.isCurrent}
+              isCompleted={step.isCompleted}
+              canNavigateTo={step.canNavigateTo}
+              isApplicable={step.isApplicable}
+              fixedText={step.fixedText}
+            />
+          ))}
+        </List>
+      </nav>
+    </Card>
   );
 };
 
 ProgressIndicator.propTypes = {
-  title: PropTypes.string,
-  submission: Types.Submission,
+  progressIndicatorTitle: PropTypes.string.isRequired,
+  formTitle: PropTypes.string.isRequired,
   steps: PropTypes.arrayOf(
     PropTypes.shape({
-      url: PropTypes.string.isRequired,
-      uuid: PropTypes.string.isRequired,
-      index: PropTypes.number.isRequired,
-      slug: PropTypes.string.isRequired,
-      href: PropTypes.string,
+      uuid: PropTypes.string,
+      slug: PropTypes.string,
+      to: PropTypes.string,
       formDefinition: PropTypes.string.isRequired,
+      isCompleted: PropTypes.bool,
       isApplicable: PropTypes.bool,
+      isCurrent: PropTypes.bool,
+      canNavigateTo: PropTypes.bool,
+      fixedText: PropTypes.string,
     })
   ).isRequired,
-  submissionAllowed: PropTypes.oneOf(Object.values(SUBMISSION_ALLOWED)).isRequired,
-  completed: PropTypes.bool,
-  hideNonApplicableSteps: PropTypes.bool,
+  ariaMobileIconLabel: PropTypes.string.isRequired,
+  accessibleToggleStepsLabel: PropTypes.string.isRequired,
+  sticky: PropTypes.bool,
 };
 
 export default ProgressIndicator;
