+++ src/components/ProgressIndicator/utils.js
